fixes after reviewing
