+++ src/client/components/events/EventPage.tsx
@@ -55,9 +55,7 @@ const EventPage = () => {
     );
   }
 
-  const categories = [...event.categories].sort(
-    (a, b) => a.displayOrder - b.displayOrder
-  );
+  const categories = event.categories;
 
   return (
     <div className="max-w-4xl mx-auto px-2 sm:px-4">
