+++ src/server/routes/eventRoutes.ts
@@ -1,5 +1,7 @@
 import { Router } from "express";
-import { authenticateToken } from "../middleware/auth";
+import { authenticateToken, authorizeAdmin } from "../middleware/auth";
+import { validateUUIDParam } from "../middleware/validation";
+import { handleValidationErrors } from "../middleware/errorHandler";
 import {
   getEvents,
   getEventBySlug,
@@ -24,13 +26,13 @@ router.get("/:slug", getEventBySlug);
 router.post("/picks", authenticateToken, submitPick);
 router.get("/:slug/my-picks", authenticateToken, getMyPicks);
 
-// Admin (authenticateToken + admin check in controller)
-router.post("/admin/events", authenticateToken, createEvent);
-router.put("/admin/events/:id", authenticateToken, updateEvent);
-router.post("/admin/categories", authenticateToken, upsertCategory);
-router.delete("/admin/categories/:id", authenticateToken, deleteCategory);
-router.post("/admin/nominees", authenticateToken, upsertNominee);
-router.delete("/admin/nominees/:id", authenticateToken, deleteNominee);
-router.put("/admin/nominees/:id/winner", authenticateToken, setWinner);
+// Admin (authenticateToken verifies JWT, authorizeAdmin checks role, validateUUIDParam validates :id)
+router.post("/admin/events", authenticateToken, authorizeAdmin, createEvent);
+router.put("/admin/events/:id", authenticateToken, authorizeAdmin, ...validateUUIDParam, handleValidationErrors, updateEvent);
+router.post("/admin/categories", authenticateToken, authorizeAdmin, upsertCategory);
+router.delete("/admin/categories/:id", authenticateToken, authorizeAdmin, ...validateUUIDParam, handleValidationErrors, deleteCategory);
+router.post("/admin/nominees", authenticateToken, authorizeAdmin, upsertNominee);
+router.delete("/admin/nominees/:id", authenticateToken, authorizeAdmin, ...validateUUIDParam, handleValidationErrors, deleteNominee);
+router.put("/admin/nominees/:id/winner", authenticateToken, authorizeAdmin, ...validateUUIDParam, handleValidationErrors, setWinner);
 
 export default router;
